FS: rename index_facade to index, add deprecation warnings
